+++ src/main/kotlin/mbuhot/eskotlin/query/term/Terms.kt
@@ -6,8 +6,8 @@ package mbuhot.eskotlin.query.term
 
 import mbuhot.eskotlin.query.QueryData
 import mbuhot.eskotlin.query.initQuery
-import org.elasticsearch.index.query.TermsQueryBuilder
-import org.elasticsearch.indices.TermsLookup
+import org.opensearch.index.query.TermsQueryBuilder
+import org.opensearch.indices.TermsLookup
 
 class TermsBlock {
     class TermsData(