+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.8.0
  
 - Upgrade to new java and latest gradle
+- change namespace from 'org.elasticsearch' to 'org.opensearch'
 
 # 0.7.0